Launch file changes for abby arm joint setter.
Made arm.launch have an argument whether to connect to the robot
(default true).
Made arm_joint_setter.launch not be a streaming commander, but just
command the rviz sim. (This should be reverted.)
Created sim_visualizer.launch to view the arm in rviz but not
connect to the IRC5.