Added a package to control abby's gripper with an arduino.
Untested.
Uses rosserial to add some gpio to control the gripper solenoid.
The arduino runs a ros service called gripper that can take open,
close, and check commands. Open and close are self-explanatory, and
check just returns the current state (open or closed) of the gripper.
All commands return a success/failure value (always returns success)
and the current state (open or closed) of the gripper.