Added gripper to the model.
Created a xacro macro for the gripper and added it to the model.
I may have to change the offset from the body to the gripper jaws,
which looks too large. All other measures should be correct. Effort
and velocity limits are bogus, but required by the model dtd.

Also altered show_model so that it doesn't try to launch the kinect.