Completed minimally-functional battery analyzer.
Added a new message PowerState to the cRIO receiver to publish
voltage measurements.

Added a new battery analyzer script, written in Python, to run
down the battery using one of three behaviors and output the
voltage data to the screen or a csv file stored in ~/cwru_battery_analyzer.

The battery analyzer can idle, oscillate back and forth in z rotations, or
swing the arm back and forth. Right now, it is hard-coded to idle and to
output results to file.

The battery analyzer automatically shuts down if the battery voltage
drops below 21.5 volts.