Started object manipulation controller.
Controller can call the segmentation service and send the results
to the collision map interface.

Controller then takes graspable objects from collision map and
attempts to each one to the object manipulator to pick up, but
this does not yet work. I need to stop and reorganize launch files
because they're getting ridiculous.